+++ internal/archiver/archiver_windows_test.go
@@ -1,0 +1,28 @@
+// +build windows
+
+package archiver
+
+import (
+	"os"
+	"testing"
+)
+
+type wrappedFileInfo struct {
+	os.FileInfo
+	mode os.FileMode
+}
+
+func (fi wrappedFileInfo) Mode() os.FileMode {
+	return fi.mode
+}
+
+// wrapFileInfo returns a new os.FileInfo with the mode, owner, and group fields changed.
+func wrapFileInfo(t testing.TB, fi os.FileInfo, mode os.FileMode, uid, gid uint) os.FileInfo {
+	// wrap the os.FileInfo and return the modified mode, uid and gid are ignored on Windows
+	res := wrappedFileInfo{
+		FileInfo: fi,
+		mode:     mode,
+	}
+
+	return res
+}
